requirements.txt
# ==============================================================================
# OG-LANS 项目依赖 (2026 Publication Ready)
# 
# 安装方式:
#   pip install -r requirements.txt
# 
# 核心依赖:
#   - unsloth: DPO 训练框架
#   - trl: 偏好学习库
#   - transformers: 模型加载
#   - networkx: 本体图构建
# ==============================================================================

absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==4.3.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
einops==0.7.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.110.3
fastjsonschema==2.21.2
filelock==3.20.2
# flash-attn==2.7.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
gast==0.7.0
grpcio==1.76.0
h11==0.16.0
hf_transfer==0.1.9
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
Markdown==3.10
markdown-it-py==2.2.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
modelscope==1.13.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openai==2.14.0
optimum==2.1.0
oss2==2.19.1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.0
pip==25.3
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==13.7.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.4.2
scikit-multilearn==0.2.0
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.3.0
# torch==2.6.0+cu118
torchao==0.15.0
# torchaudio==2.6.0+cu118
# torchvision==0.21.0+cu118
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
triton==3.2.0
trl==0.24.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==1.0.3
tzdata==2025.3
unsloth==2026.1.4
unsloth_zoo==2026.1.4
urllib3==2.6.2
uvicorn==0.29.0
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
wheel==0.45.1
xxhash==3.6.0
yapf==0.43.0
yarg==0.1.9
yarl==1.22.0
dirtyjson>=1.0.8